Stockagent/pyproject.toml
[project]
name = "stockagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.8.1",
    "langchain>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "mcp-use>=1.3.1",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.63",
]
